Flashy surprise board
An idea I came up with, just to experiment a bit. This might be a bit big for some screens, but with just a bit of simple JavaScript, I was able to make a board where you have to see what surprise you get. Simple but effective. My first original project.